+++ core/management/commands/data_migration.py
@@ -19,14 +19,14 @@ from psycopg2.extras import DictCursor
 from core import recref_settings
 from core.helper import iter_utils
 from core.models import CofkUnionResource, CofkUnionComment, CofkLookupDocumentType, CofkUnionRelationshipType, \
-    CofkUnionImage, CofkUnionOrgType, CofkUnionRoleCategory, CofkUnionSubject, Iso639LanguageCode
+    CofkUnionImage, CofkUnionOrgType, CofkUnionRoleCategory, CofkUnionSubject, Iso639LanguageCode, CofkLookupCatalogue
 from institution.models import CofkUnionInstitution
 from location.models import CofkUnionLocation
 from login.models import CofkUser
 from manifestation.models import CofkUnionManifestation
 from person.models import CofkUnionPerson, SEQ_NAME_COFKUNIONPERSION__IPERSON_ID
 from publication.models import CofkUnionPublication
-from uploader.models import CofkCollectStatus, CofkLookupCatalogue, CofkCollectUpload
+from uploader.models import CofkCollectStatus, CofkCollectUpload
 from work import models as work_models
 from work.models import CofkUnionWork, CofkUnionQueryableWork
 from audit.models import CofkUnionAuditLiteral, CofkUnionAuditRelationship
